+++ src/components/RickRoll.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
-  export let title: string;
-  export let url: string;
+  export let data:
+    | { type: "youtube"; title: string; url: string }
+    | { type: "video"; src: string };
   export let roll: string;
   export let chance: string;
 </script>
@@ -8,14 +9,24 @@
 <div>
   <h1 class="text-4xl font-bold">You Lucky!</h1>
   <hr class="my-4" />
-  <iframe
-    class="mx-auto aspect-video w-full"
-    src="https://www.youtube.com/embed/{url}?autoplay=1"
-    {title}
-    frameborder="0"
-    allow="autoplay; encrypted-media; picture-in-picture"
-    allowfullscreen
-  />
+  {#if data.type === "youtube"}
+    <iframe
+      class="mx-auto aspect-video w-full"
+      src="https://www.youtube.com/embed/{data.url}?autoplay=1"
+      title={data.title}
+      frameborder="0"
+      allow="autoplay; encrypted-media; picture-in-picture"
+      allowfullscreen
+    />
+  {:else}
+    <!-- svelte-ignore a11y-media-has-caption -->
+    <video
+      class="mx-auto aspect-video w-full"
+      src={data.src}
+      autoplay
+      controls
+    />
+  {/if}
   <p class="mt-4 text-xl font-semibold sm:text-2xl">
     There is {chance}% Chance you will be {roll} and you get it!
   </p>